+++ pico/src/fir_filter.cpp
@@ -6,7 +6,7 @@ using namespace std;
 
 // 32x oversampling, lowpass 67200Hz
 
-static const int16_t coefficients[] = {
+static const int32_t coefficients[] = {
 	-0.071383729317764918 * 0x7fff,
 	0.055470186813273974 * 0x7fff,
 	0.304697734025869083 * 0x7fff,
@@ -23,13 +23,13 @@ reg(), coeff(coefficients)
 
 // Update returns false if no new data is available.
 // Update returns true if new data has been placed in out.
-bool FIRFilter::Update(const int16_t in, int16_t *out) {
-	uint32_t tmp;
+bool FIRFilter::Update(const int32_t in, int32_t *out) {
+	int32_t tmp;
 	// Shift in new value
 	this->reg.Update(in);
 
 	// Apply the filter
-	tmp = ShiftReg<int16_t, 7>::Convolute<int16_t,int16_t, 7, 7>(this->reg, this->coeff);
+	tmp = ShiftReg<int32_t, 7>::Convolute<int32_t,int32_t, 7, 7>(this->reg, this->coeff);
 	*out = tmp >> 15;
 	return true;
 }
